FIX: internal server error detection only on auto-submit

+++ gerritautopush.sh
@@ -270,7 +270,7 @@ function push {
 		fi
 		
 		# something failed; we need to analyze what went wrong
-		if [ `cat $TMPFILE | grep -e 'remote rejected.*Internal server error' | wc -l` > 0 ]; then
+		if [ `cat $TMPFILE | grep -e 'remote rejected.*Internal server error' | wc -l` > 0 ] && [ $AUTO_SUBMIT == "true" ]; then
 			rm -f $TMPFILE
 			echo "WARNING: Internal server error occured; trying to push again after 3 seconds"
 			sleep 3
